+++ test/unit/wizards/connectedap.test.ts
@@ -18,7 +18,6 @@ import {
 } from '../../../src/foundation.js';
 import {
   createConnectedApWizard,
-  editConnectedAp104Wizard,
   editConnectedApWizard,
 } from '../../../src/wizards/connectedap.js';
 import { ListItemBase } from '@material/mwc-list/mwc-list-item-base';
@@ -127,117 +126,6 @@ describe('Wizards for SCL element ConnectedAP', () => {
       expect(
         oldAddress.querySelector<Element>('P[type="IP"]')?.textContent
       ).to.equal('192.168.210.111');
-      expect(
-        newAddress.querySelector<Element>('P[type="IP"]')?.textContent
-      ).to.equal('192.168.210.158');
-    });
-
-    it('adds type restrictions with selected option type restriction', async () => {
-      (<Checkbox>(
-        element.wizardUI.shadowRoot?.querySelector('#typeRestriction')
-      )).checked = true;
-      await element.requestUpdate();
-
-      primaryAction.click();
-      await element.requestUpdate();
-
-      const complexAction = <ComplexAction>actionEvent.args[0][0].detail.action;
-
-      const oldAddress = <Element>(
-        (<Delete>complexAction.actions[0]).old.element
-      );
-      const newAddress = <Element>(
-        (<Create>complexAction.actions[1]).new.element
-      );
-
-      const oldIP = oldAddress.querySelector<Element>('P[type="IP"]');
-      const newIP = newAddress.querySelector<Element>('P[type="IP"]');
-
-      expect(
-        oldIP?.getAttributeNS(
-          'http://www.w3.org/2001/XMLSchema-instance',
-          'type'
-        )
-      ).to.not.exist;
-      expect(
-        newIP?.getAttributeNS(
-          'http://www.w3.org/2001/XMLSchema-instance',
-          'type'
-        )
-      ).to.exist;
-    });
-  });
-
-  describe('include an edit wizard for the 104 plugin that', () => {
-    beforeEach(async () => {
-      doc = await fetch('/test/testfiles/104-protocol.scd')
-        .then(response => response.text())
-        .then(str => new DOMParser().parseFromString(str, 'application/xml'));
-
-      const wizard = editConnectedAp104Wizard(doc.querySelector('SubNetwork[type="104"] > ConnectedAP[apName="AP2"]')!);
-      element.workflow.push(() => wizard);
-      await element.requestUpdate();
-
-      inputs = Array.from(element.wizardUI.inputs);
-
-      primaryAction = <HTMLElement>(
-        element.wizardUI.dialog?.querySelector(
-          'mwc-button[slot="primaryAction"]'
-        )
-      );
-    });
-
-    it('does not edit any P element with unchanged wizard inputs', async () => {
-      primaryAction.click();
-      await element.requestUpdate();
-      expect(actionEvent).to.not.have.been.called;
-    });
-
-    it('does not trigger a complex editor action to update P elements(s) when not all fields are filled in', async () => {
-      input = <WizardTextField>inputs.find(input => input.label === 'IP');
-      input.value = '';
-      await input.requestUpdate();
-
-      primaryAction.click();
-      await element.requestUpdate();
-
-      expect(actionEvent).to.not.have.been.called;
-    });
-
-    it('triggers a complex action as combination of delete and create with prior existing Address field', async () => {
-      input = <WizardTextField>inputs.find(input => input.label === 'IP');
-      input.value = '192.128.0.1';
-      await input.requestUpdate();
-
-      primaryAction.click();
-      await element.requestUpdate();
-
-      const complexAction = <ComplexAction>actionEvent.args[0][0].detail.action;
-      expect(complexAction.actions).to.have.lengthOf(2);
-      expect(complexAction.actions[0]).to.satisfy(isDelete);
-      expect(complexAction.actions[1]).to.satisfy(isCreate);
-    });
-
-    it('properly updates a P element of type IP', async () => {
-      input = <WizardTextField>inputs.find(input => input.label === 'IP');
-      input.value = '192.168.210.158';
-      await input.requestUpdate();
-
-      primaryAction.click();
-      await element.requestUpdate();
-
-      const complexAction = <ComplexAction>actionEvent.args[0][0].detail.action;
-
-      const oldAddress = <Element>(
-        (<Delete>complexAction.actions[0]).old.element
-      );
-      const newAddress = <Element>(
-        (<Create>complexAction.actions[1]).new.element
-      );
-
-      expect(
-        oldAddress.querySelector<Element>('P[type="IP"]')?.textContent
-      ).to.equal('192.128.0.11');
       expect(
         newAddress.querySelector<Element>('P[type="IP"]')?.textContent
       ).to.equal('192.168.210.158');
